feat(keycloak): Add keycloak theme, capture university, add ORCid integration (#971)
We build a custom keycloak theme with keycloakify (which is react based), which gets packaged (like most Keycloak themes) into a .jar file. We then put that jar file into a Docker image which serves just as a directory to store the image, and copies it out into a volume that is also mounted on keycloak in the right place for keycloak to get the theme. Keycloak does indeed gets the theme and uses it as the login theme.

We also enable the declarative_user_profile feature of Keycloak, and use it to capture the user's University / Organisation.

We also add ORCid keycloak integration and capture the ORCid where possible