Rename instances to organisms and more (#1209)
This PR makes two changes:

it renames instances in helm Values to organisms, in line with our nomenclature. This was a backlog item.
it renames the organisms in the default Values.yaml to defaultOrganisms. It only uses this if organisms is unset. Without this, if we make a new instance and set organisms appropriately, the dummy-organism, etc., gets added in too. It would be nice to do this with helm templates but I couldn't make that work.