feat (keycloak): add email verification, password reset to keycloak (#1320)
Adds email functionality to keycloak. To get the password to keycloak, the config processor is updated to additionally substitute from environmental variables (which can contain secrets). Also we rename "loculusRealm" to "loculus" - in part as the emails surface this name.

We are currently using Theo's Mailjet account - we probably can for a while - it has a limit of 15k emails per month, split with some other projects.