feat(website/backend): Button for sequence revocation
Add a revocation button to the bottom of the sequence details page if the sequence is from one of my groups
Make it redirect to review page
Make review page display revocations awaiting approval
Make approval endpoint accept revocations awaiting approval
Bring in react-toastify and use it to display error messages here:
(We have what is IMO a bad pattern of using "ManagedErrorFeedback", which requires us to store in react state the current error condition. The react-toastify approach is much simpler and just lets us say toast(errorMessage) without managing any state. I think we should gradually transition to that. (Or to a version of our own feedback element that works in the same way). I don't want to add another example of the IMO not great pattern that we'll have to adapt later.)