aider: Added keycloak-deployment.yaml and updated database service and deployment to be conditional on keycloak database credentials not being set.
# Aider chat conversation:

USER: add `kubernetes/loculus/templates/keycloak-deployment.yaml`

ASSISTANT: To add the `kubernetes/loculus/templates/keycloak-deployment.yaml` file, here is the diff:

```diff
--- /dev/null
+++ kubernetes/loculus/templates/key
^C KeyboardInterrupt

USER: now update the database service and database deployment to be conditional on these not being set

ASSISTANT: To conditionally include the Keycloak database service and deployment based on whether the Keycloak database credentials are set, we need to update both `kubernetes/loculus/templates/keycloak-database-service.yaml` and `kubernetes/loculus/templates/keycloak-database-deployment.yaml`. Here are the diffs for these changes:

For `kubernetes/loculus/templates/keycloak-database-service.yaml`:

```diff
--- kubernetes/loculus/templates/keycloak-database-service.yaml
+++ kubernetes/loculus/templates/keycloak-database-service.yaml
@@ ... @@
+{{- if not .Values.keycloakDatabase.urlSealedSecret }}
 apiVersion: v1
 kind: Service
 metadata:
   name: loculus-keycloak-database-service
 spec:
   type: ClusterIP
   selector:
     app: loculus
     component: keycloak-database
   ports:
     - port: 5432
+{{- end }}
```

For `kubernetes/loculus/templates/keycloak-database-deployment.yaml`:

```diff
--- kubernetes/loculus/templates/keycloak-database-deployment.yaml
+++ kubernetes/loculus/templates/keycloak-database-deployment.yaml
@@ ... @@
+{{- if not .Values.keycloakDatabase.urlSealedSecret }}
 apiVersion: apps/v1
 kind: Deployment
 metadata:
   name: loculus-keycloak-database
   annotations:
     argocd.argoproj.io/sync-options: Replace=true
     reloader.stakater.com/auto: "true"
 spec:
   replicas: 1
   selector:
     matchLabels:
       app: loculus
       component: keycloak-database
   template:
     metadata:
       labels:
         app: loculus
         component: keycloak-database
     spec:
       containers:
         - name: loculus-keycloak-database
           image: postgres:latest
           ports:
             - containerPort: 5432
           env:
             - name: POSTGRES_USER
               value: "postgres"
             - name: POSTGRES_PASSWORD
               value: "unsecure"
             - name: POSTGRES_DB
               value: "keycloak"
+{{- end }}
```

These changes wrap the entire content of both files in a conditional block that checks if `.Values.keycloakDatabase.urlSealedSecret` is not set. If it is set, the service and deployment for the Keycloak database will not be created, assuming that an external database is being used.