feat(website): Add a header option to metadata (#1636)
This PR adds an optional `header` field to the values.yaml file. The default value is undefined and in this case all metadata fields without a header will be grouped together in one table and shown at the start. Otherwise metadata with the same header will be grouped and shown together under that header. 
Headers are added to the MetaData table by creating a `toHeaderMap` function which maps headers to the metadata with that corresponding header.