feat(website): return organism early when found, increase robustness to failing lapis (#2052)
Resolves #2051

Rather than waiting for all LAPISes to say whether they have a sequence, we just need to wait for the first that confirms that a sequence is in that organism. This speeds up our seq details queries and also makes them resilient against unrelated organisms being down.

One should be able to notice that seq details is snappier like this.