fix(ingest): also retry reverse complement when guessing segment
As we rely on nextclade to identify segments, switching on `--retry-reverse-complement` allows us to deal with reverse complemented sequences.

For CCHF, there are 25 extra sequences that we can identify with this option enabled.

We need to be **very careful** with this fix as it can change grouping - something that ingest might not be able to deal with in the current state.