feat(kubernetes): make public runtime config urls of the website configurable
For GenSpectrum, we need to be able to set the public URLs of the services that the website accesses.
Actually this is a generalized case of the "codespaceName" feature,
so my approach was to pull out the `$.Values.codespaceName` parameter from the Helm chart into `deploy.py`.
The website URLs can be set in values.yaml via
```
website:
  runtetimeConfig:
    public:
      backendUrl: ...
      # where %organism% is a placeholder for the organisms in the config
      lapisUrlTemplate: https://%organism%.my.lapis.org
      keycloakUrl: ...
```