chore(helm): move earliestReleaseDate to _common-metadata.tpl
Followup to #3380

Hard coded metadata fields are configured in _common-metadata.tpl as they are not something to be configured per instance via values.yaml

Disabling per organism isn't easy for common-metadata fields, so let's remove the ability to disable, earliest release is something that can always be calculated so no need to disable it. Let's only make configurable what truly needs to be configurable.