Publish files on sequence approval (#3967)
resolves #3931 

Adds a new function to fetch all the File IDs for files that were
approved, and then iterates through them and sets them to public using
the tag `public=true`.

For this to work, the S3 needs to be configured with ah policy:

```json
{
  "Version": "2012-10-17",
  "Statement": [
    {
      "Effect":     "Allow",
      "Action":     "s3:GetObject",
      "Resource":    "arn:aws:s3:::loculus-preview-private/*",
      "Principal":   "*",
      "Condition": {  "StringEquals": {"s3:ExistingObjectTag/public": "true" } }
    }
  ]
}
```

I already applied this policy to `loculus-preview-private`.

🚀 Preview: Add `preview` label to enable

---------

Co-authored-by: Chaoran Chen <mail@chaoran-chen.de>
Co-authored-by: GitHub Action <action@github.com>