feat(website): Strip `\r` and `\n` on paste into text field (#3985)
resolves #3655.

N.B. This does not resolve the underlying issue
https://github.com/loculus-project/loculus/issues/3878

It is unlikely we want linebreaks in singleline fields - it is quite
easily to accidentally paste them and then they can cause confusion.
This becomes very common due to
https://github.com/loculus-project/loculus/issues/3878. Here we strip
these characters from single line fields as a workaround. For now this
only addresses singleline text fields.

I tested that this works


### PR Checklist
- [x] The implemented feature is covered by appropriate, automated
tests.
- [x] Any manual testing that has been done is documented (i.e. what
exactly was tested?)

🚀 Preview: https://pastefix.loculus.org