feat(deployment): timestamp pods on deployment to force restart of all pods on every deployment (#4034)
resolves #3170

This adds timestamp annotations to all pods that basically describe the
time on which helm was run to make a deployment. The effect of this
should be that all pods have to restart when a new deployment is made,
because this value will change. This should ensure that config-only
updates are still made.

Testing:
- Tested that the product still works
- Tested that I can see the annotations in the pod

In terms of whether it fixes the issue - it should but IMO we can wait
and see

🚀 Preview: https://timetamp-now.loculus.org