fix(ingest): handle revocations correctly in call_loculus (#4042)
resolves
https://github.com/loculus-project/loculus/pull/4040#issuecomment-2823260215

### Screenshot

If we now revoke 2 sequences that were created by the insdc_ingest_user
we get another error. Original metadata is empty for revocations and
without this the call_loculus script maps the revoked sequences to
insdc_accession=NULL. We throw an error if multiple sequences map to the
same insdc_accession - it also good to make sure this are mapped
correctly so we never try to revise a revoked sequence.

### PR Checklist
- [ ] All necessary documentation has been adapted.
- [ ] The implemented feature is covered by appropriate, automated
tests.
- [ ] Any manual testing that has been done is documented (i.e. what
exactly was tested?)

🚀 Preview: https://original-metadata-endpoin.loculus.org