fix(backend,ingest): handle revoked sequences in /get-original-metadata, they have null metadata (#4040)
resolves #4036 

co-authored by @anna-parker

Fix backend to deal with `null` original metadata.

Now we get this for revoked sequences in `/get-original-metadata` (we
add the `isRevocation` field to the output):

```
{"accession":"LOC_000M2HN","version":2,"submitter":"superuser","originalMetadata":null, "isRevocation": True}
```

Ingest also works: we remove the key sort and additionally handle
revocations separately (as they do not contain the INSDC accession in
the originalMetadata as it is empty)

We now emit isRevocation in `/get-original-metadata` to allow ingest to
handle revocations properly.

### Testing

- Unit test added that reproduces the bug and no longer does so with the
fix
- Manual in preview: Revoke a sequence, rerun ingest without problems
- Manual in preview: Revoke two sequences, rerun ingest without problems
(Anya noticed this failed and she fixed it)

### Ideas for future

- Nice to have: Integration test for reingest with a revoked sequence

### PR Checklist
- ~[ ] All necessary documentation has been adapted.~
- [x] The implemented feature is covered by appropriate, automated
tests.
- [x] Any manual testing that has been done is documented (i.e. what
exactly was tested?)

🚀 Preview: https://fix-null-metadata.loculus.org

---------

Co-authored-by: Anna (Anya) Parker <50943381+anna-parker@users.noreply.github.com>