file sharing: Add integration tests (#4025)
resolves #3990

- make MinIO run when deploying locally
- Add a test that uploads some single submission data
- Add a test that uploads bulk data
- Add now `BulkSubmissionPage` page to playwright tests.

This also adds a new 'internal endpoint' s3 config setting. I ran into
an issue with running the backend in k3s and using 'localhost' to access
S3, which only works from outside the cluster, but caused connectivity
issues from the backend Pod. Now there are two settings for the
endpoint, one 'internal' which the backend will use when talking to S3
inside the cluster, and one external which will be used for generated
links.

### PR Checklist
- [ ] All necessary documentation has been adapted.
- [x] The implemented feature is covered by appropriate, automated
tests.
- [x] Any manual testing that has been done is documented (i.e. what
exactly was tested?)

🚀 Preview: https://fs-integr-test.loculus.org