fix(ingest): use nextclade sort for non-segmented, take top-score when using nextclade align (#4008)
1. ensure when processing alignment that highest scoring alignment is
chosen

Currently we assume a sequence will only ever align to one segment and
do not drop lower scoring results. (Use a script for this as it
otherwise gets too complicated - this is nice as I then can remove
tsv-utils from the dependencies)

This also adds a test (adds a small (mutated) segment of the M segment
to an S segment in this tests and ensures that the highest scoring
alignment - the S alignment is still chosen)

2. Allow non-segmented organisms to use nextclade-sort (this is useful
for e.g. rsv)

resolves #

### Screenshot

### PR Checklist
- [ ] All necessary documentation has been adapted.
- [x] The implemented feature is covered by appropriate, automated
tests.
- [ ] Any manual testing that has been done is documented (i.e. what
exactly was tested?)

🚀 Preview: https://fix-ingest-alignment.loculus.org