feat(ingest): notify on potential suppression (#4075)
resolves https://github.com/loculus-project/loculus/issues/2885

Send a notification on slack if re-ingested sequences do not include
insdc-accessions that were previously ingested (and are not in a
blocked, error, revoked state).

### Screenshot

### PR Checklist
- [x] All necessary documentation has been adapted.
- [x] Any manual testing that has been done is documented (i.e. what
exactly was tested?)
Test:
1. ingest from ebola zaire and set db to persistent
2. switch to ingest from ebola sudan
3. ensure receive notification:

![image](https://github.com/user-attachments/assets/0eb6a9b4-d6d4-4378-90c1-d1b79ec10533)
admittedly its not very easy to search this message - but I assume we
will generally not receive a notification of 3k sequences that are
missing and I would need to give ingest the slack token to upload files
4. revoke some sequences (e.g. ebola zaire: OK572988.1), re-ingest
ensure not notified about these sequences

![image](https://github.com/user-attachments/assets/09e6895c-cb71-49e7-a939-b274d7df1e28)

=> I am now reverting the persistent DB changes to stop us getting
notifications every 30min on slack.

🚀 Preview: https://notify-on-suppression.loculus.org