chore(github): remove special codespace configuration (#4086)
Some time ago I made a special configuration that was intended to start
a backend and a frontend in codespaces. No one is using this at the
moment, and it is very slow, and we don't even know if it still works.
Codespaces are now used as part of GitHub's Agentic CoPilot and so we
should remove this as it makes it very slow to start a codespace at all.
We can always add it back if desired.

🚀 Preview: Add `preview` label to enable