feat(ingest): fix revoke-and-regroup cronjob bug, add revocation comment (#2451)
### Summary

1. Add description of why sequence has been revoked if revoked due to
grouping changes
2. Only revoke a sequence due to a regrouping if there is some
uncertainty on how to regroup, e.g. if due to new sequences a previous
grouping must be split up or two groups have to be joined. If a new
segment can be added to an existing group (i.e. was missing before) just
add it to the group using a revision
3. Small bug fixes: give revoke_and_regroup cron-job proper permissions
to actually be able to do its job
4. Small bug fixes: use the latest `jointAccession` for comparing -
otherwise if a new segment is added we can get into a revision loop.

### PR Checklist
Use the new [CCHF nextclade
dataset](https://github.com/nextstrain/nextclade_data/pull/265) for
testing which improves alignment, resulting in many M segments that were
previously lost now being aligned and ingested.

- ingest using current dataset
- set db to persistent
- ingest and processing using new dataset
- run revoke_and_regroup cronjob
- check results

### Screenshots
- Message is sent with sequences that should be revoked and regrouped:

![image](https://github.com/user-attachments/assets/5510e6ec-88f3-46f9-8678-bcf4d7b95138)
- Message after sequence has been revoked is clear:

![image](https://github.com/user-attachments/assets/c5002f38-c400-47c1-8869-2ed046cc1fa6)

🚀 Preview: https://ingest-add-revocation-com.loculus.org