feat(website): limit max number of displayed autocomplete options (#4170)
The auto-completion of the search is crashing my browser tab if there
are too many options. As a solution, this PR proposes to only show a
limited number of options. I observed a problem in a field with ~8000
values while it worked well for ~700 values, so I chose a cutoff of
1000.