feat(prepro): add nextclade sort option to check if sequence aligns better to another organism (#4123)
resolves #

### Summary

Adds new option to prepro config: `require_nextclade_sort_match = true`,
when this is set before aligning to the reference run nextclade sort.
(This uses the default minimizer produced by nextclade for that
nextclade server). Asserts that the top scoring dataset is the dataset
we use for aligning that sequence/segment. If not we add an error to the
processed output.

If there is no match when running nextclade sort we do not throw an
error but add a warning. (The sequence will then be aligned and if this
fails an error will be thrown)

### Screenshot

![image](https://github.com/user-attachments/assets/e60583b3-761f-47f1-acc4-225e58b1ee40)


### PR Checklist
- [ ] All necessary documentation has been adapted.
- [ ] The implemented feature is covered by appropriate, automated
tests.
- [ ] Any manual testing that has been done is documented (i.e. what
exactly was tested?)

🚀 Preview: https://add-nextclade-sort-to-pre.loculus.org

---------

Co-authored-by: Cornelius Roemer <cornelius.roemer@gmail.com>