fix(backend): fix disabling of S3 (#4169)
The backend cannot start at the moment if S3 is disabled because it
still needs to initialize an `S3Service` and the service is creating an
S3 client while being initialized, for which it needs S3 connection
data. This PR fixes the problem by offering two S3Service
implementations: one real implementation that is used when S3 is enabled
and another implementation that is used when S3 is disabled. The latter
will always initialize but only throw an error when its functions are
being called.