feat!(deposition): make an api endpoint for ingest (#4091)
resolves #

Ingest currently directly queries the ena-deposition schema to not
reingest sequences we submitted - here I add an endpoint so that ingest
does not need to have database credentials.

Endpoint can be viewed via portforwarding:
```
kubectl port-forward $POD -n $NAMESPACE 5000:5000
```
 it is running at port http://localhost:5000/submitted.

### What this adds 

1. Uses [FastAPI](https://fastapi.tiangolo.com/) (enables fast, async
requests and is served via ASGI web server:
[uvicorn](https://www.uvicorn.org/)) create a simple `/submitted`
endpoint for the ena-deposition pod that can be used by ingest to query
for a list of all INSDC assembly and biosample accessions submitted by
Loculus, endpoint returns:
```
{"status":"ok","insdcAccessions":["", "", ..],"biosampleAccessions":["SAMExx", ...]}
```
2. Simplifies logic in ingest and removes DB credentials to instead use
the new endpoint
3. Modifies the values.yaml - adds all ena-related fields to one section
and adds a new field `enaGithubTestUrl` that can be used to specify
which github URL should be used to upload sequences from for testing.

### What needs to be changed

I tidied up the `values.yaml` a bit, now all the enaDeposition-related
params should go under:
```
enaDeposition: 
  submitToEnaProduction: false
  enaDbName: Loculus
  enaUniqueSuffix: Loculus
  enaIsBroker: False
```


### PR Checklist
- [x] All necessary documentation has been adapted => schema has been
updated
- [x] The implemented feature is covered by appropriate, automated
tests.
- [x] Any manual testing that has been done is documented (i.e. what
exactly was tested?)
-> ingest and the ena-deposition pod were run locally and code was
tested, also run on preview, sadly ENA dev's env is down so I had to log
into the DB to add some test examples but I could see that this was
successful. I also verified that the api server will be killed if
threads in other jobs are killed -> this is important to restart the pod
and for kubernetes logging


![image](https://github.com/user-attachments/assets/3b9904f1-af6b-4ba5-b337-ac3417c0fba4)
and is used by ingest:
```
12:26:10    DEBUG (get_loculus_depositions.py:  82) - Assembly accessions to filter out: []
12:26:10    DEBUG (get_loculus_depositions.py:  85) - Biosample accessions to filter out: ['SAMEA131597971']
```

🚀 Preview: https://make-ena-deposition-pod-f.loculus.org