[BREAKING] file sharing: use canonical links (#4064)
**BREAKING** - the `website.runtimeConfig.public` section has been moved
to the top level, i.e. is now directly at `public`.

---

resolves #3969 


This PR adds a new endpoint to the frontend to get a link to a file with
the pattern:
`$websiteUrl/seq/$accession.$version/$fileCategory/$fileName`. There is
a new config setting (`filesSharing.outputFileUrlType`) that allows
configuring whether this kind of link should be output by
`get-released-data` or whether to use a direct link to S3.

The website URL uses the filename, and so files downloaded are also
using the filename.

An integration test was added to test this.

**Manual testing**: I've manually checked that the link looks ok, and
I've also tested `website` vs `s3` URLs in the preview to check that the
Helm Chart works.

We plan to use the new `fileSharing` top level config object later on
for more file sharing related settings, such as the expiry time of a pre
signed URL, or garbage collector related things.

### Screenshot


https://github.com/user-attachments/assets/6bde9882-868c-4cab-8aa3-efbdf42e5201

### PR Checklist
- [x] All necessary documentation has been adapted. -> the schema was
expanded.
- [x] The implemented feature is covered by appropriate, automated
tests.
- [x] Any manual testing that has been done is documented (i.e. what
exactly was tested?)

🚀 Preview: https://file-sharing-canonical-ur.loculus.org