refactor(backend): disable S3 client in a different way (#4184)
This was Codex's suggestion for how to allow disabling S3 without
creating a Noop interface. It's no big deal either way - but it does
seem decent to my untrained eye. I was also just interested to test
Codex on the backend (it's a challenge because it doesn't have docker
for the testContainers, but can still be useful).

- drop `S3ServiceNoop`
- update `S3ServiceImpl` to always exist and lazily create clients
- adjust tests for the new behaviour

@chaoran-chen  etc., np if you think this isn't a good approach
(integration test failure is file-sharing tests)

🚀 Preview: https://codex-refactor-s3-client.loculus.org

---------

Co-authored-by: Chaoran Chen <mail@chaoran-chen.de>