fix(website): Check for empty LAPIS responses and raise error (#4224)
resolves #4223 

LAPIS resonds with an `OK` response also for empty data, and that caused
issues. Now we're checking if at least one data row is returned, and
otherwise an error is raised (making the Promise rejected).

Note on redirects: for URLs without a version, a redirect is returned,
so that is just passed through as OK - we "kick the can down to road"

### Manual testing

I tested that I can DL sequence data on seqs where I couldn't (reliably)
before. I also tested a link without a version part and that worked too.
This is the URL I used during my tests:
https://main.loculus.org/seq/LOC_00004U7.1 (but with localhost)

### PR Checklist
- ~~All necessary documentation has been adapted.~~ (n/A)
- ~~The implemented feature is covered by appropriate, automated
tests.~~ (postponed: #4225)
- [x] Any manual testing that has been done is documented (i.e. what
exactly was tested?)

🚀 Preview: https://seq-dl-fix.loculus.org