feat(config): reduce the resources reserved for the config processor (#4233)
resolves #

### Screenshot
We discussed last week that pod cpu reservation is the max of all
containers incl. the init containers. Setting resources means the pod
must have at least this much cpu/mem available to be able to start
leading to some previews not being scheduled. We checked last week that
we often only need a fraction of this.

### PR Checklist
- [ ] All necessary documentation has been adapted.
- [ ] The implemented feature is covered by appropriate, automated
tests.
- [ ] Any manual testing that has been done is documented (i.e. what
exactly was tested?)

🚀 Preview: https://update-resources-config-p.loculus.org