feat(website): add metadata fields and organisms to seqsets table (#4248)
Now SeqSets are getting use we fairly urgently need to make them look
less basic as they will be many people's first entrypoint to
Pathoplexus. In particular we need to make sure they credit submitters
effectively without further clicks. This implements that in a fairly
hacky way where we lazily query LAPIS once we get a list of accessions.
It's also not tested for now - we need to bring the SeqSets over to the
integration tests but I don't think this is blocking and can't take that
on now.

<img width="1117" alt="image"
src="https://github.com/user-attachments/assets/34aa7ac0-62c7-4002-ac87-d383a309bc4d"
/>


## Summary

This PR enhances the seqsets table by:
1. Making the it display metadata fields configurable via Helm values
2. Adding an organism column that shows user-friendly display names

## Changes

- Added `fieldsToDisplay` configuration under `seqSets` in `values.yaml`
with default fields:
  - Country (`geoLocCountry`)
  - Collection Date (`sampleCollectionDate`) 
  - Authors (`authors`)
- Updated Helm templates to pass this configuration to the website
- Made the seq-sets table query LAPIS for these metadata when it
receives the list of accessions

🚀 Preview: https://seqset-metadata.loculus.org

---------

Co-authored-by: Cornelius Roemer <cornelius.roemer@gmail.com>